pyproject.toml
[project]
name = "MaybecleverPaperFilter"
version = "0.1.0"
description = "A tool for a lazy postdoctoral researcher. XD"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.11",
    "ollama>=0.4.7",
    "python-telegram-bot>=21.11.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "telegraph>=2.2.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
